+++ installers/Installer.hs
@@ -9,8 +9,6 @@ import qualified System.IO as IO
 
 import qualified MacInstaller
 import qualified WindowsInstaller
-import qualified Data.Text.IO as T
-import qualified Data.Text as T
 import           Data.Yaml                 (decodeFileThrow)
 
 import           Types
@@ -33,10 +31,9 @@ main = do
     GenInstaller -> do
         genSignedInstaller (oOS options') options'
     BuildkiteCrossWin -> do
-      fullVersion <- getDaedalusVersion "../package.json"
-      ver <- T.strip <$> T.readFile "version" -- TODO
-      let fullName = packageFileName Win64 (oCluster options') fullVersion (oBackend options') ver (oBuildJob options')
+      fullVersion <- getAppVersion "../package.json"
       installerConfig <- decodeFileThrow "installer-config.json"
+      let fullName = packageFileName (uglyName installerConfig) Win64 (oCluster options') fullVersion (oBackend options') (oBuildJob options') (oBuildCounter options')
       WindowsInstaller.writeInstallerNSIS fullName fullVersion installerConfig options' (oCluster options')
       WindowsInstaller.writeUninstallerNSIS fullVersion installerConfig
 -- | The contract of `genSignedInstaller` is not to produce unsigned installer binaries.