+++ installers/common/Util.hs
@@ -4,12 +4,8 @@ module Util where
 
 import Data.Text (Text)
 import System.Directory (listDirectory, withCurrentDirectory, removeDirectory, removeFile, doesDirectoryExist)
-import Turtle (export)
 import Data.Aeson (Value, Value(Object, String), encodeFile, decodeFileStrict')
-import qualified Data.HashMap.Strict as HM
-
-import Config (Options(..), Backend(..))
-import Types (fromBuildJob, clusterNetwork)
+import qualified Data.Aeson.KeyMap as HM
 
 windowsRemoveDirectoryRecursive :: FilePath -> IO ()
 windowsRemoveDirectoryRecursive path = do
@@ -21,21 +17,6 @@ windowsRemoveDirectoryRecursive path = do
         removeDirectory path
     else do
         removeFile path
-
--- | Sets the many environment variables required for the
--- "yarn package" build.
--- When updating this, check that all variables are baked in with both
--- webpack.config.js files.
-exportBuildVars :: Options -> Text -> IO ()
-exportBuildVars Options{oBackend, oBuildJob, oCluster} backendVersion = do
-    mapM_ (uncurry export)
-        [ ("API", apiName oBackend)
-        , ("API_VERSION", backendVersion)
-        , ("BUILD_NUMBER", maybe "" fromBuildJob oBuildJob)
-        , ("NETWORK", clusterNetwork oCluster)
-        ]
-    where
-        apiName (Cardano _) = "ada"
 
 rewritePackageJson :: FilePath -> Text -> IO ()
 rewritePackageJson path name = do