+++ source/renderer/app/api/index.ts
@@ -9,7 +9,7 @@ export type Api = {
 
 export const setupApi = (isTest: boolean): Api => ({
   ada: new AdaApi(isTest, {
-    hostname: 'localhost',
+    hostname: '127.0.0.1', // “localhost” breaks under new electron, which prefers ::1 (IPv6)
     port: 8090,
     ca: Uint8Array.from([]),
     key: Uint8Array.from([]),